+++ janus/__init__.py
@@ -8,7 +8,22 @@ from heapq import heappop, heappush
 from queue import Empty as SyncQueueEmpty
 from queue import Full as SyncQueueFull
 from time import monotonic
-from typing import Any, Callable, Generic, Optional, Protocol, TypeVar
+from typing import Callable, Generic, Optional, Protocol, TypeVar
+
+if sys.version_info >= (3, 13):
+    from asyncio import QueueShutDown as AsyncQueueShutDown
+    from queue import ShutDown as SyncQueueShutDown
+else:
+    class QueueShutDown(Exception):
+        pass
+
+    AsyncQueueShutDown = QueueShutDown
+
+    class ShutDown(Exception):
+        pass
+
+    SyncQueueShutDown = ShutDown
+
 
 __version__ = "1.2.0"
 __all__ = (
@@ -18,9 +33,11 @@ __all__ = (
     "SyncQueue",
     "SyncQueueEmpty",
     "SyncQueueFull",
+    "SyncQueueShutDown",
     "AsyncQueue",
     "AsyncQueueEmpty",
     "AsyncQueueFull",
+    "AsyncQueueShutDown",
     "BaseQueue",
 )
 
@@ -51,6 +68,8 @@ class BaseQueue(Protocol[T]):
 
     def get_nowait(self) -> T: ...
 
+    def shutdown(self, immediate: bool = False) -> None: ...
+
 
 class SyncQueue(BaseQueue[T], Protocol[T]):
 
@@ -77,6 +96,7 @@ class Queue(Generic[T]):
             self._loop = asyncio.get_running_loop()
 
         self._maxsize = maxsize
+        self._is_shutdown = False
 
         self._init(maxsize)
 
@@ -101,8 +121,7 @@ class Queue(Generic[T]):
         self._async_tasks_done = asyncio.Condition(self._async_mutex)
         self._async_tasks_done_waiting = 0
 
-        self._closing = False
-        self._pending: deque[asyncio.Future[Any]] = deque()
+        self._pending: deque[asyncio.Future[None]] = deque()
 
         self._sync_queue = _SyncQueueProxy(self)
         self._async_queue = _AsyncQueueProxy(self)
@@ -119,39 +138,72 @@ class Queue(Generic[T]):
             raise RuntimeError(f"{self!r} is bound to a different event loop")
         return loop
 
-    def close(self) -> None:
+    def shutdown(self, immediate: bool = False) -> None:
+        """Shut-down the queue, making queue gets and puts raise an exception.
+
+        By default, gets will only raise once the queue is empty. Set
+        'immediate' to True to make gets raise immediately instead.
+
+        All blocked callers of put() and get() will be unblocked. If
+        'immediate', a task is marked as done for each item remaining in
+        the queue, which may unblock callers of join().
+
+        The raise exception is SyncQueueShutDown for sync api and AsyncQueueShutDown
+        for async one.
+        """
         with self._sync_mutex:
-            self._closing = True
-            for fut in self._pending:
-                fut.cancel()
-            if self._async_tasks_done_waiting:
-                # unblocks all async_q.join()
-                self._notify_async(self._async_tasks_done.notify_all)
-            if self._sync_tasks_done_waiting:
-                self._sync_tasks_done.notify_all()  # unblocks all sync_q.join()
+            self._is_shutdown = True
+            if immediate:
+                while self._qsize():
+                    self._get()
+                    if self._unfinished_tasks > 0:
+                        self._unfinished_tasks -= 1
+                # release all blocked threads in `join()`
+                if self._sync_tasks_done_waiting:
+                    self._sync_tasks_done.notify_all()
+                if self._async_tasks_done_waiting:
+                    self._notify_async(self._async_tasks_done.notify_all)
+            # All getters need to re-check queue-empty to raise ShutDown
+            if self._sync_not_empty_waiting:
+                self._sync_not_empty.notify_all()
+            if self._sync_not_full_waiting:
+                self._sync_not_full.notify_all()
+            if self._async_not_empty_waiting:
+                self._notify_async(self._async_not_empty.notify_all)
+            if self._async_not_full_waiting:
+                self._notify_async(self._async_not_full.notify_all)
+
+    def close(self) -> None:
+        """Close the queue.
+
+        The method is a shortcut for .shutdown(immediate=True)
+        """
+        self.shutdown(immediate=True)
 
     async def wait_closed(self) -> None:
+        """Wait for finishing all pending activities"""
         # should be called from loop after close().
         # Nobody should put/get at this point,
         # so lock acquiring is not required
-        if not self._closing:
+        if not self._is_shutdown:
             raise RuntimeError("Waiting for non-closed queue")
-        # give execution chances for the task-done callbacks
+        # give a chance for the task-done callbacks
         # of async tasks created inside
-        # _make_async_not_empty_notifier, _make_async_not_full_notifier
-        # methods.
+        # _notify_async()
+        # methods to be executed.
         await asyncio.sleep(0)
         if not self._pending:
             return
         await asyncio.wait(self._pending)
 
     async def aclose(self) -> None:
+        """Shutdown the queue and wait for actual shutting down"""
         self.close()
         await self.wait_closed()
 
     @property
     def closed(self) -> bool:
-        return self._closing and not self._pending
+        return self._is_shutdown and not self._pending
 
     @property
     def maxsize(self) -> int:
@@ -208,10 +260,6 @@ class Queue(Generic[T]):
             return
         loop.call_soon_threadsafe(self._setup_async_notifier, loop, method)
 
-    def _check_closing(self) -> None:
-        if self._closing:
-            raise RuntimeError("Operation on the closed queue is forbidden")
-
 
 class _SyncQueueProxy(SyncQueue[T]):
     """Create a queue object with a given maximum size.
@@ -245,7 +293,6 @@ class _SyncQueueProxy(SyncQueue[T]):
         placed in the queue.
         """
         parent = self._parent
-        parent._check_closing()
         with parent._sync_tasks_done:
             unfinished = parent._unfinished_tasks - 1
             if unfinished <= 0:
@@ -267,7 +314,6 @@ class _SyncQueueProxy(SyncQueue[T]):
         When the count of unfinished tasks drops to zero, join() unblocks.
         """
         parent = self._parent
-        parent._check_closing()
         with parent._sync_tasks_done:
             while parent._unfinished_tasks:
                 parent._sync_tasks_done_waiting += 1
@@ -275,7 +321,6 @@ class _SyncQueueProxy(SyncQueue[T]):
                     parent._sync_tasks_done.wait()
                 finally:
                     parent._sync_tasks_done_waiting -= 1
-                parent._check_closing()
 
     def qsize(self) -> int:
         """Return the approximate size of the queue (not reliable!)."""
@@ -322,8 +367,9 @@ class _SyncQueueProxy(SyncQueue[T]):
         is ignored in that case).
         """
         parent = self._parent
-        parent._check_closing()
         with parent._sync_not_full:
+            if parent._is_shutdown:
+                raise SyncQueueShutDown
             if parent._maxsize > 0:
                 if not block:
                     if parent._qsize() >= parent._maxsize:
@@ -335,6 +381,8 @@ class _SyncQueueProxy(SyncQueue[T]):
                             parent._sync_not_full.wait()
                         finally:
                             parent._sync_not_full_waiting -= 1
+                        if parent._is_shutdown:
+                            raise SyncQueueShutDown
                 elif timeout < 0:
                     raise ValueError("'timeout' must be a non-negative number")
                 else:
@@ -348,6 +396,8 @@ class _SyncQueueProxy(SyncQueue[T]):
                             parent._sync_not_full.wait(remaining)
                         finally:
                             parent._sync_not_full_waiting -= 1
+                        if parent._is_shutdown:
+                            raise SyncQueueShutDown
             parent._put_internal(item)
             if parent._sync_not_empty_waiting:
                 parent._sync_not_empty.notify()
@@ -366,8 +416,9 @@ class _SyncQueueProxy(SyncQueue[T]):
         in that case).
         """
         parent = self._parent
-        parent._check_closing()
         with parent._sync_not_empty:
+            if parent._is_shutdown and not parent._qsize():
+                raise SyncQueueShutDown
             if not block:
                 if not parent._qsize():
                     raise SyncQueueEmpty
@@ -378,6 +429,8 @@ class _SyncQueueProxy(SyncQueue[T]):
                         parent._sync_not_empty.wait()
                     finally:
                         parent._sync_not_empty_waiting -= 1
+                    if parent._is_shutdown and not parent._qsize():
+                        raise SyncQueueShutDown
             elif timeout < 0:
                 raise ValueError("'timeout' must be a non-negative number")
             else:
@@ -391,6 +444,8 @@ class _SyncQueueProxy(SyncQueue[T]):
                         parent._sync_not_empty.wait(remaining)
                     finally:
                         parent._sync_not_empty_waiting -= 1
+                    if parent._is_shutdown and not parent._qsize():
+                        raise SyncQueueShutDown
             item = parent._get()
             if parent._sync_not_full_waiting:
                 parent._sync_not_full.notify()
@@ -413,6 +468,21 @@ class _SyncQueueProxy(SyncQueue[T]):
         raise the Empty exception.
         """
         return self.get(block=False)
+
+    def shutdown(self, immediate: bool = False) -> None:
+        """Shut-down the queue, making queue gets and puts raise an exception.
+
+        By default, gets will only raise once the queue is empty. Set
+        'immediate' to True to make gets raise immediately instead.
+
+        All blocked callers of put() and get() will be unblocked. If
+        'immediate', a task is marked as done for each item remaining in
+        the queue, which may unblock callers of join().
+
+        The raise exception is SyncQueueShutDown for sync api and AsyncQueueShutDown
+        for async one.
+        """
+        self._parent.shutdown(immediate)
 
 
 class _AsyncQueueProxy(AsyncQueue[T]):
@@ -471,9 +541,10 @@ class _AsyncQueueProxy(AsyncQueue[T]):
         This method is a coroutine.
         """
         parent = self._parent
-        parent._check_closing()
         async with parent._async_not_full:
             with parent._sync_mutex:
+                if parent._is_shutdown:
+                    raise AsyncQueueShutDown
                 parent._get_loop()  # check the event loop
                 while 0 < parent._maxsize <= parent._qsize():
                     parent._async_not_full_waiting += 1
@@ -483,6 +554,8 @@ class _AsyncQueueProxy(AsyncQueue[T]):
                     finally:
                         parent._sync_mutex.acquire()
                         parent._async_not_full_waiting -= 1
+                    if parent._is_shutdown:
+                        raise AsyncQueueShutDown
 
                 parent._put_internal(item)
                 if parent._async_not_empty_waiting:
@@ -496,8 +569,10 @@ class _AsyncQueueProxy(AsyncQueue[T]):
         If no free slot is immediately available, raise QueueFull.
         """
         parent = self._parent
-        parent._check_closing()
         with parent._sync_mutex:
+            if parent._is_shutdown:
+                raise AsyncQueueShutDown
+
             parent._get_loop()
             if 0 < parent._maxsize <= parent._qsize():
                 raise AsyncQueueFull
@@ -516,9 +591,10 @@ class _AsyncQueueProxy(AsyncQueue[T]):
         This method is a coroutine.
         """
         parent = self._parent
-        parent._check_closing()
         async with parent._async_not_empty:
             with parent._sync_mutex:
+                if parent._is_shutdown and not parent._qsize():
+                    raise AsyncQueueShutDown
                 parent._get_loop()  # check the event loop
                 while not parent._qsize():
                     parent._async_not_empty_waiting += 1
@@ -528,6 +604,8 @@ class _AsyncQueueProxy(AsyncQueue[T]):
                     finally:
                         parent._sync_mutex.acquire()
                         parent._async_not_empty_waiting -= 1
+                    if parent._is_shutdown and not parent._qsize():
+                        raise AsyncQueueShutDown
 
                 item = parent._get()
                 if parent._async_not_full_waiting:
@@ -542,8 +620,9 @@ class _AsyncQueueProxy(AsyncQueue[T]):
         Return an item if one is immediately available, else raise QueueEmpty.
         """
         parent = self._parent
-        parent._check_closing()
         with parent._sync_mutex:
+            if parent._is_shutdown and not parent._qsize():
+                raise AsyncQueueShutDown
             if not parent._qsize():
                 raise AsyncQueueEmpty
 
@@ -570,7 +649,6 @@ class _AsyncQueueProxy(AsyncQueue[T]):
         the queue.
         """
         parent = self._parent
-        parent._check_closing()
         with parent._sync_tasks_done:
             if parent._unfinished_tasks <= 0:
                 raise ValueError("task_done() called too many times")
@@ -590,7 +668,6 @@ class _AsyncQueueProxy(AsyncQueue[T]):
         When the count of unfinished tasks drops to zero, join() unblocks.
         """
         parent = self._parent
-        parent._check_closing()
         async with parent._async_tasks_done:
             with parent._sync_mutex:
                 parent._get_loop()  # check the event loop
@@ -602,7 +679,21 @@ class _AsyncQueueProxy(AsyncQueue[T]):
                     finally:
                         parent._sync_mutex.acquire()
                         parent._async_tasks_done_waiting -= 1
-                    parent._check_closing()
+
+    def shutdown(self, immediate: bool = False) -> None:
+        """Shut-down the queue, making queue gets and puts raise an exception.
+
+        By default, gets will only raise once the queue is empty. Set
+        'immediate' to True to make gets raise immediately instead.
+
+        All blocked callers of put() and get() will be unblocked. If
+        'immediate', a task is marked as done for each item remaining in
+        the queue, which may unblock callers of join().
+
+        The raise exception is SyncQueueShutDown for sync api and AsyncQueueShutDown
+        for async one.
+        """
+        self._parent.shutdown(immediate)
 
 
 class PriorityQueue(Queue[T]):